+++ docs/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Changed
+
+-   Made min version of eulertools a mandatory field
+
 ### Added
 
 -   Added a way to allow passing the problem name with an id