+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [3.0.0] - 2024-05-20
+
 ### Changed
 
 -   Made min version of eulertools a mandatory field
@@ -137,7 +139,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/eulertools/compare/v2.1.0...main
+[Unreleased]: https://github.com/spapanik/eulertools/compare/v3.0.0...main
+[3.0.0]: https://github.com/spapanik/eulertools/compare/v2.1.0...v3.0.0
 [2.1.0]: https://github.com/spapanik/eulertools/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/spapanik/eulertools/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/spapanik/eulertools/compare/v0.10.0...v1.0.0